GUI/requirements.txt
PyQt6==6.3.1
pyserial==3.5
matplotlib==3.5.3
